Extract pointer management to Maps::PointersChannel
