+++ app/views/campaigns/_current_map.html.erb
@@ -2,7 +2,7 @@
   <%= content_tag :div,
     class: "flex-1 flex flex-row h-full",
     data: {
-      controller: "map map--tokens #{"map--fog" if campaign.current_map.fog_enabled?}",
+      controller: "map map--tokens map--pointers #{"map--fog" if campaign.current_map.fog_enabled?}",
       "map-id": campaign.current_map.id,
       action: "dragover->map--tokens#dragOver"
     } do %>
@@ -20,8 +20,8 @@
       data: {
         controller: "css-var #{"sync-values" if admin}",
         "map-id": campaign.current_map.id,
-        target: "map.image map--tokens.image #{"map--fog.image" if campaign.current_map.fog_enabled?}",
-        action: "dragover->map--tokens#dragOverMap drop->map--tokens#dropOnMap mousedown->map#pointTo #{"mousedown->map--fog#startDrawingFog" if campaign.current_map.fog_enabled? && admin} mousedown->map#moveMap",
+        target: "map.image map--tokens.image map--pointers.image #{"map--fog.image" if campaign.current_map.fog_enabled?}",
+        action: "dragover->map--tokens#dragOverMap drop->map--tokens#dropOnMap mousedown->map--pointers#pointTo #{"mousedown->map--fog#startDrawingFog" if campaign.current_map.fog_enabled? && admin} mousedown->map#moveMap",
         x: campaign.current_map.x,
         y: campaign.current_map.y,
         zoom: campaign.current_map.zoom,
@@ -39,7 +39,7 @@
         "sync-values-keys": "x y"
       } do %>
       <%= content_tag :div, "", class: "current-map__image", style: background_image(campaign.current_map.image) %>
-      <div class="current-map__token-container" data-target="map.tokenContainer map--tokens.container">
+      <div class="current-map__token-container" data-target="map--tokens.container map--pointers.container">
         <%= render campaign.current_map.tokens.where(stashed: false) %>
       </div>
